+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Bump `hasql` dependencies
 * Introduce `destroyPgSqlStore` function
+* Add realtime event stream observer
 
 ## 0.1.0.1